当前层级字段相关分析器优化
