+++ demo_2.py
@@ -2,10 +2,8 @@
 解析 MySQL 建表语句
 """
 
-from metasequoia_sql import SQLParser, SQLType
-from metasequoia_sql.analyzer.current_level_common_column import CurrentGroupByClauseUsedQuoteColumn
-from scripts.demo_sql.sql_basic_tutorial import *
-import dataclasses
+from metasequoia_sql import SQLParser
+from metasequoia_sql.analyzer.toolkit.current_level_column_analyzer import CurrentGroupByClauseUsedQuoteColumn
 
 if __name__ == "__main__":
     # statement = SQLParser.parse_select_statement(SBT_CH02_33)
